# _openapi_rag-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "roxie-rag-service",
    "description": "汽车 DTC 故障码检索增强生成服务。\n\n四路 async 并行召回 + RRF 融合 + 规则重排 + Cross-Encoder 精排 + 断崖截断。\n\n## 接口概览\n\n| 接口 | 说明 |\n| --- | --- |\n| `POST /api/v1/diagnoses` | DTC 故障诊断检索(主接口) |\n| `POST /api/v1/context/extract` | 故障上下文槽位提取 |\n| `POST /api/v1/dtc-groups` | DTC 候选记录分组(独立计算) |\n| `POST /api/v1/candidate-causes/rank` | 候选原因排序(独立计算) |\n| `POST /api/v1/dtc-cause-cards` | DTC 故障组原因处置卡片聚合 |\n| `GET /api/v1/tools` | Roxie 工具发现:列出全部工具 |\n| `GET /api/v1/tools/{tool_name}` | Roxie 工具发现:单个工具完整定义 |\n| `POST /api/v1/tools/{tool_name}/invoke` | 五工具统一调用入口 |\n| `POST /api/v1/tool-services/dtc-context` | 强类型:DTC 背景上下文解析 |\n| `POST /api/v1/tool-services/dtc-grouping` | 强类型:多 DTC 关联分组 |\n| `POST /api/v1/tool-services/cause-ranking` | 强类型:候选原因检索/排序 |\n| `POST /api/v1/tool-services/diagnostic-planning` | 强类型:诊断检查项/计划 |\n| `POST /api/v1/tool-services/repair-planning` | 强类型:维修计划生成 |\n| `GET /api/v1/live` | 存活探针 |\n| `GET /api/v1/ready` | 就绪探针(MySQL) |\n| `GET /api/v1/health` | 深度健康检查(五组件) |\n",
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/v1/tools/{tool_name}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "工具发现：获取单个工具完整定义",
        "operationId": "get_tool_api_v1_tools__tool_name__get",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "工具不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_name}/invoke": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "统一调用：执行工具",
        "operationId": "invoke_tool_api_v1_tools__tool_name__invoke_post",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolInvokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolInvokeResponse"
                }
              }
            }
          },
          "404": {
            "description": "工具不存在"
          },
          "400": {
            "description": "arguments 不符合工具 input_schema"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolDetail": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "tool_kind": {
            "type": "string",
            "title": "Tool Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          },
          "output_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Schema"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "tool_kind",
          "description",
          "skill_name",
          "input_schema",
          "output_schema"
        ],
        "title": "ToolDetail"
      },
      "ToolDetailResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ToolDetail"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "data",
          "trace_id"
        ],
        "title": "ToolDetailResponse"
      },
      "ToolInvokeRequest": {
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments",
            "description": "工具入参对象，须满足工具详情返回的 input_schema"
          }
        },
        "type": "object",
        "title": "ToolInvokeRequest"
      },
      "ToolInvokeResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "trace_id"
        ],
        "title": "ToolInvokeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}